Move cryptofunction service to km
