+++ libs/common/src/auth/services/webauthn-login/webauthn-login-prf-key.service.spec.ts
@@ -1,6 +1,6 @@
 import { mock, MockProxy } from "jest-mock-extended";
 
-import { CryptoFunctionService } from "../../../platform/abstractions/crypto-function.service";
+import { CryptoFunctionService } from "../../../key-management/crypto/abstractions/crypto-function.service";
 
 import { WebAuthnLoginPrfKeyService } from "./webauthn-login-prf-key.service";
