+++ libs/common/src/key-management/crypto/services/bulk-encrypt.service.impementation.spec.ts
@@ -2,11 +2,11 @@ import { mock, MockProxy } from "jest-mock-extended";
 import * as rxjs from "rxjs";
 
 import { ServerConfig } from "../../../platform/abstractions/config/server-config";
-import { CryptoFunctionService } from "../../../platform/abstractions/crypto-function.service";
 import { LogService } from "../../../platform/abstractions/log.service";
 import { Decryptable } from "../../../platform/interfaces/decryptable.interface";
 import { InitializerMetadata } from "../../../platform/interfaces/initializer-metadata.interface";
 import { SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-crypto-key";
+import { CryptoFunctionService } from "../abstractions/crypto-function.service";
 import { buildSetConfigMessage } from "../types/worker-command.type";
 
 import { BulkEncryptServiceImplementation } from "./bulk-encrypt.service.implementation";
