+++ libs/common/src/key-management/crypto/services/encrypt.worker.ts
@@ -9,7 +9,7 @@ import { SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-cr
 import { ConsoleLogService } from "../../../platform/services/console-log.service";
 import { ContainerService } from "../../../platform/services/container.service";
 import { getClassInitializer } from "../../../platform/services/cryptography/get-class-initializer";
-import { WebCryptoFunctionService } from "../../../platform/services/web-crypto-function.service";
+import { WebCryptoFunctionService } from "./web-crypto-function.service";
 import {
   DECRYPT_COMMAND,
   SET_CONFIG_COMMAND,
