+++ libs/common/src/key-management/crypto/services/web-crypto-function.service.ts
@@ -1,11 +1,14 @@
 import * as argon2 from "argon2-browser";
 import * as forge from "node-forge";
 
-import { Utils } from "../../platform/misc/utils";
-import { CsprngArray } from "../../types/csprng";
+import { Utils } from "../../../platform/misc/utils";
+import {
+  CbcDecryptParameters,
+  EcbDecryptParameters,
+} from "../../../platform/models/domain/decrypt-parameters";
+import { SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-crypto-key";
+import { CsprngArray } from "../../../types/csprng";
 import { CryptoFunctionService } from "../abstractions/crypto-function.service";
-import { CbcDecryptParameters, EcbDecryptParameters } from "../models/domain/decrypt-parameters";
-import { SymmetricCryptoKey } from "../models/domain/symmetric-crypto-key";
 
 export class WebCryptoFunctionService implements CryptoFunctionService {
   private crypto: Crypto;
