+++ libs/common/src/platform/services/key-generation.service.spec.ts
@@ -2,8 +2,8 @@ import { mock } from "jest-mock-extended";
 
 import { PBKDF2KdfConfig, Argon2KdfConfig } from "@bitwarden/key-management";
 
+import { CryptoFunctionService } from "../../key-management/crypto/abstractions/crypto-function.service";
 import { CsprngArray } from "../../types/csprng";
-import { CryptoFunctionService } from "../abstractions/crypto-function.service";
 
 import { KeyGenerationService } from "./key-generation.service";
