+++ libs/importer/src/importers/lastpass/access/services/crypto-utils.ts
@@ -1,4 +1,4 @@
-import { CryptoFunctionService } from "@bitwarden/common/platform/abstractions/crypto-function.service";
+import { CryptoFunctionService } from "@bitwarden/common/key-management/crypto/abstractions/crypto-function.service";
 import { Utils } from "@bitwarden/common/platform/misc/utils";
 
 export class CryptoUtils {
