+++ libs/node/src/services/node-crypto-function.service.ts
@@ -2,7 +2,7 @@ import * as crypto from "crypto";
 
 import * as forge from "node-forge";
 
-import { CryptoFunctionService } from "@bitwarden/common/platform/abstractions/crypto-function.service";
+import { CryptoFunctionService } from "@bitwarden/common/key-management/crypto/abstractions/crypto-function.service";
 import { Utils } from "@bitwarden/common/platform/misc/utils";
 import {
   CbcDecryptParameters,
